Remove LayoutCommandsExt & add bevy_ui dsl impl
LayoutCommandsExt was silly, and I couldn't make the traits work anyway.
Now, the methods are on `LayoutCommands` rather than
`LayoutCommandsExt`. At the expense of having to call a `.lyout()`
method to convert a `cmds` into `LayoutCommands`, we don't have any more
trait head-aches.

It still isn't super composable, but it's already pretty good.