Fix panic on certain cyclic rules + better errors
There was a todo! in Rule::inside. I discovered I already had a way to
deal with this in the LeafRule::inside, just the `unwrap` method
implemented on `Size` was not generic enough.

I'm surprised I didn't catch that.

I also reworded some error messages & use Flow (aliased to Axis) instead
of a &str in the Why type.