Spawn terminal nodes as leaf nodes
When using the `spawn` DSL statement, LayoutDsl would insert anything as
a container node.

This is a problem, since `spawn` can be used to spawn terminal nodes.
This would result in an unavoidable error when using `spawn` with
LayoutDsl.

What we now do is check that we are neither row or column, and if so,
spawn as Node::Box ourselves.