Set color of ui image bundles
The NodeBundle default background color is transparent.

This is a problem when spawning UiImage, since then its alpha is
multiplied by the background color and itself becomes invisible.

So we add a BackgroundColor field to ImageBundle and set it to white.

This allows spawning images through the `image` method and the `spawn`
DSL statement, rather than `spawn_ui`.