Fix bad offset when using bevy_ui borders
Fixes #38.

The problem stems from `position: absolute` being offset by the parent's
border while it shouldn't.

Somehow, `border` influences children's "absolute" position, but not
padding. Which means we can't easily work around this by setting a
negative padding equivalent to the border size.

The solution I opted for was to spawn an additional child to the node
with border, and put the border on that child instead of the node
itself.

Alternatives considered:

- Add a `BorderOffset` component to entities with a border, then add a
  system in ComputeLayoutSet running after the layout computation that
  adjusts the offset of children with `BorderOffset` component.
- Try fiddling with other style fields. But no cigar