Use Children(1.5) as default layout rule
Children(1.0) could be confusing in situations where all chldren has the
same corss size. If the children each have the same uniform color, it
becomes difficult to notice distinct nodes, and could lead to confusion.

Ideally, the default — if not ideal — should be the least confusing, in
order to enable users to quickly proceed to fixing up stuff.

Using Children(1.5) enables adding "gap" between children, which is
easier to parse in some situations.