Add the layout LayoutDsl method
`.layout` accepts a &str argument and parses it according to the
specification described in design_docs/debug.md.

It is more succint that calling the corresponding distribution and
alignment methods, but also, since it is going to be used in the debug
overlay, it makes the users familiar with it.