Add screen space handling + proper window scaling
… To layout debug overlay.

This makes it possible to use it with cuicui_layout_bevy_sprite, which
will be inverted on the Y axis in compariison with cuicui_layout_bevy_ui

We also use the window scale instead of hardcoding the scale in order to
properly evaluate line width. This makes the overlay pixel-perfect on
every plateform.