Use a marker component for compute_content_size
Since only leaf nodes can be content-sized, we mark them specifically
and use the marker to discriminate between entities we update the
Node of, and entities that are not going to be updated.

Now, we don't need a buffer to hold updates and can directly update in
place the content.

Fixes #35