Add a KDL-based file format
This defines a file format to write dsl trees as independent files.

Currently, the format is based on KDL, but the plan is to migrate into a
fully independent parser (to avoid any kind of allocation & make it
easier to copy/paste code declared with the dsl! macro)

Missing features:

- A &ref system where we can pass references to the file format
- An AssetLoader to integrate with bevy and enable hot-reloading
- A dsl! macro-like syntax instead of KDL
- Good error reporting pointing at the exact position in file of errors
  & using miette or ariadne for nice messages
- An attribute macro to automatically implement ParseDsl based on impl
  blocks.
- Any handling of `Handle` or loading `Handle`s declared in the file
  format
- Better handling of generic methods.
- Implementations for existing DSLs