Hot reload chirp scenes
We weren't listening to asset changes for chirp files, which broke hot
reloading. There were also some tricky bits to get right.

Hot reloading didn't work because we kept around the InstanceId of the
old Handle<Scene>, as a result, we would spawn dumb useless entities.

The trick was to delete the instance in the ChirpInstances resource, to
force the spawner to re-spawn a new scene with a new instance id.

Still buggy:

- Doesn't handle at all multiple instances of the same scene.
- Entities are inserted half-hazardly in parent/root. Ideally we track
  where the seed was and insert in the children list at the same spot.