Add text handling to UiDsl
This allows using the chirp loader for the bevypunk demo.

Since TextFlags and TextLayoutInfo are not registered components, I
I couldn't add them to the chirp scene, so I removed them from the UiDsl
TextBundle and add them after-the-fact in
text_fixup::add_text_components.

I also adjust the parser to accept quoted text in argument position, and
to pass forward unquoted text.