Fix inset calculation overflow
In some circumstances the inseting computation would result in an
overflow, I've no idea why, especially since those are operations on
_signed_ integers.

This is probably a very bad fix, since `increment` is always 1 or -1,
it implies the computation is applied several millions time. But it
fixes a major limitation of hot reloading, so it goes in.

Fixes #84