Add multiple argument template & template extras
- The `template!() (method) {Child}` syntax is now supported and tested
- This includes deeply nested setups, and proper argument/parameter
  handling.
- Now, several arguments can be passed to templates, furthermore, nested
  template calls also work as expected.

This required a complete rewrite of the interpreter. Now instead of
being tied to the parser, the parser produces an AST, that is then
interpreted.

This also changes the way method arguments are passed to
ParseDsl::method. Now, instead of requiring re-parsing, we pass the
already-split input.

This is a very naive AST, with a lot of allocations. A better AST
implementation is drafted in `design_docs/ast.md`.