Check for uninitialized headers at AST build time
Previously, we did it whenever the header reference was dropped.
However, the header could be dropped before being initialized in
legitimate situations, such as parsing errors.

Now, we manage a stack of uninitialized headers, and pop it whenever it
is initialized. If we call AstBuilder::build while the stack is not
empty, we panic, because we know our AST is invalid

We also add back `const` on some pre-existing functions

Fixes #102